# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Diliké",
        "version": "1.0"
    },
    "paths": {
...
        "/api/login_user": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Connexion",
                "description": "Connexion et retour des données de l'utilisateur",
                "operationId": "login_user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "zouzoua@gmail.com",
                                    "password": "12345678X"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "liste des utilisateurs",
                "description": "liste de de tous les utilisateurs",
                "operationId": "users_list",
                "responses": {
                    "200": {
                        "description": "Afficher la liste de de tous les utilisateurs"
                    }
                }
            },
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Inscription",
                "description": "Inscription d'un utilisateur",
                "operationId": "users",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenoms": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "XXXXX",
                                    "prenoms": "XXXXX",
                                    "phone": "xxxx",
                                    "email": "XXXXX@gmail.com",
                                    "password": "12345678X",
                                    "password_confirmation": "12345678X"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Obtenir un utilisateur",
                "description": "Information sur un utilisateur",
                "operationId": "on_user",
                "responses": {
                    "200": {
                        "description": "Information sur un utilisateur"
                    }
                }
            },
            "put": {
                "tags": [
                    "users"
                ],
                "summary": "Mise à jour d'un utilisateur",
                "description": "Mise à jour d'un utilisateur",
                "operationId": "users_update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenoms": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Kouadio",
                                    "prenoms": "Essis",
                                    "phone": "56854215",
                                    "password": "12345678X"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Suprimer un utilisateur",
                "description": "Suprimer un utilisateur",
                "operationId": "user_delete",
                "responses": {
                    "200": {
                        "description": "Suprimer un utilisateur"
                    }
                }
            }
        }
    }
}